apply more suggestions
