+++ deepinv/models/__init__.py
@@ -15,6 +15,5 @@ from .swinir import SwinIR
 from .PDNet import PDNet_PrimalBlock, PDNet_DualBlock
 from .bm3d import BM3D
 from .equivariant import EquivariantDenoiser
-from .gmm import GaussianMixtureModel
 from .epll import EPLL
 from .patchnr import PatchNR
