+++ deepinv/models/epll.py
@@ -4,6 +4,8 @@ from deepinv.models.gmm import GaussianMixtureModel
 from deepinv.utils import patch_extractor
 from deepinv.optim.utils import conjugate_gradient
 from deepinv.utils.demo import load_torch_url
+from deepinv.physics import Denoising, GaussianNoise
+from tqdm import tqdm
 
 
 class EPLL(nn.Module):
@@ -55,26 +57,39 @@ class EPLL(nn.Module):
                     raise ValueError("Pretrained weights not found!")
                 weights = load_torch_url(url)
             self.GMM.load_parameter_dict(weights)
+        self.denoising_operator = Denoising(GaussianNoise())
 
-            self.GMM.load_parameter_dict(weights)
+    def forward(self, x, sigma, betas=None, batch_size=-1):
+        r"""
+        Calls the reconstruction for denoising
 
-    def forward(self, x):
+        :param torch.Tensor x: tensor of noisy images. Shape: batch size x ...
+        :param float sigma: noise level
+        :param list of floats betas: parameters from the half-quadratic splitting. None uses the standard choice 1/sigma_sq [1,4,8,16,32]
+        :param int batch_size: batching the patch estimations for large images. No effect on the output, but a small value reduces the memory consumption
+            but might increase the computation time. -1 for considering all patches at once.
+        """
+        return self.reconstruction(
+            x, x.clone(), sigma, self.denoising_operator, batch_size=batch_size
+        )
+
+    def negative_log_likelihood(self, x):
         r"""
         Takes patches and returns the negative log likelihood of the GMM for each patch.
 
         :param torch.Tensor x: tensor of patches of shape batch_size x number of patches per batch x patch_dimensions
         """
         B, n_patches = x.shape[0:2]
-        logpz = self.GMM.negative_log_likelihood(x.view(B * n_patches, -1))
+        logpz = self.GMM(x.view(B * n_patches, -1))
         return logpz.view(B, n_patches)
 
-    def reconstruction(self, y, x_init, sigma_sq, physics, betas=None, batch_size=-1):
+    def reconstruction(self, y, x_init, sigma, physics, betas=None, batch_size=-1):
         r"""
         Approximated half-quadratic splitting method for image reconstruction as proposed by Zoran and Weiss.
 
         :param torch.Tensor y: tensor of observations. Shape: batch size x ...
         :param torch.Tensor x_init: tensor of initializations. Shape: batch size x channels x height x width
-        :param float sigma_sq: squared noise level (acts as regularization parameter)
+        :param float sigma: noise level (acts as regularization parameter)
         :param deepinv.physics.LinearPhysics physics: Forward operator. Has to be linear. Requires physics.A and physics.A_adjoint.
         :param list of floats betas: parameters from the half-quadratic splitting. None uses the standard choice 1/sigma_sq [1,4,8,16,32]
         :param int batch_size: batching the patch estimations for large images. No effect on the output, but a small value reduces the memory consumption
@@ -82,7 +97,7 @@ class EPLL(nn.Module):
         """
         if betas is None:
             # default choice as suggested in Parameswaran et al. "Accelerating GMM-Based Patch Priors for Image Restoration: Three Ingredients for a 100× Speed-Up"
-            betas = [beta / sigma_sq for beta in [1.0, 4.0, 8.0, 16.0, 32.0]]
+            betas = [beta / sigma**2 for beta in [1.0, 4.0, 8.0, 16.0, 32.0]]
         if y.shape[0] > 1:
             # vectorization over a batch of images not implemented....
             return torch.cat(
@@ -100,7 +115,7 @@ class EPLL(nn.Module):
         x = x_init
         Aty = physics.A_adjoint(y)
         for beta in betas:
-            x = self._reconstruction_step(Aty, x, sigma_sq, beta, physics, batch_size)
+            x = self._reconstruction_step(Aty, x, sigma**2, beta, physics, batch_size)
         return x
 
     def _reconstruction_step(self, Aty, x, sigma_sq, beta, physics, batch_size):
@@ -159,3 +174,239 @@ class EPLL(nn.Module):
         op = lambda im: physics.A_adjoint(physics.A(im)) + beta * sigma_sq * im
         hat_x = conjugate_gradient(op, rhs, max_iter=1e2, tol=1e-5)
         return hat_x
+
+
+class GaussianMixtureModel(nn.Module):
+    r"""
+    Gaussian mixture model including parameter estimation.
+
+    Implements a Gaussian Mixture Model, its negative log likelihood function and an EM algorithm
+    for parameter estimation.
+
+    :param int n_components: number of components of the GMM
+    :param int dimension: data dimension
+    :param str device: gpu or cpu.
+    """
+
+    def __init__(self, n_components, dimension, device="cpu"):
+        super(GaussianMixtureModel, self).__init__()
+        self._covariance_regularization = None
+        self.n_components = n_components
+        self.dimension = dimension
+        self._weights = torch.ones((n_components,), device=device)
+        self.set_weights(self._weights)
+        self.mu = torch.zeros((n_components, dimension), device=device)
+        self._cov = 0.1 * torch.eye(dimension, device=device)[None, :, :].tile(
+            n_components, 1, 1
+        )
+        self.set_cov(self._cov)
+
+    def set_cov(self, cov):
+        r"""
+        Sets the covariance parameters to cov and maintains their log-determinants and inverses
+
+        :param torch.Tensor cov: new covariance matrices in a n_components x dimension x dimension tensor
+        """
+        self._cov = cov.to(self._cov)
+        self._logdet_cov = torch.logdet(self._cov)
+        self._cov_inv = torch.linalg.inv(self._cov)
+        if self._covariance_regularization:
+            self._cov_reg = self._cov + self._covariance_regularization * torch.eye(
+                self.dimension, device=self._cov.device
+            )[None, :, :].tile(self.n_components, 1, 1)
+            self._logdet_cov_reg = torch.logdet(self._cov_reg)
+            self._cov_inv_reg = torch.linalg.inv(self._cov_reg)
+
+    def set_cov_reg(self, reg):
+        r"""
+        Sets covariance regularization parameter for evaluating
+        Needed for EPLL.
+
+        :param float reg: covariance regularization parameter
+        """
+        self._covariance_regularization = reg
+        self._cov_reg = self._cov + self._covariance_regularization * torch.eye(
+            self.dimension, device=self._cov.device
+        )[None, :, :].tile(self.n_components, 1, 1)
+        self._logdet_cov_reg = torch.logdet(self._cov_reg)
+        self._cov_inv_reg = torch.linalg.inv(self._cov_reg)
+
+    def get_cov(self):
+        r"""
+        get method for covariances
+        """
+        return self._cov.clone()
+
+    def get_cov_inv_reg(self):
+        r"""
+        get method for covariances
+        """
+        return self._cov_inv_reg.clone()
+
+    def set_weights(self, weights):
+        r"""
+        sets weight parameter while ensuring non-negativity and summation to one
+
+        :param torch.Tensor weights: non-zero weight tensor of size n_components with non-negative entries
+        """
+        assert torch.min(weights) >= 0.0
+        assert torch.sum(weights) > 0.0
+        self._weights = (weights / torch.sum(weights)).to(self._weights)
+
+    def get_weights(self):
+        r"""
+        get method for weights
+        """
+        return self._weights.clone()
+
+    def get_parameter_dict(self):
+        r"""
+        Creates dict with parameters
+        """
+        parameters = {
+            "weights": self._weights.clone().cpu(),
+            "mu": self.mu.cpu(),
+            "cov": self._cov.clone().cpu(),
+        }
+        return parameters
+
+    def load_parameter_dict(self, parameter_dict):
+        r"""
+        Loads dict with parameters
+
+        :param dict parameter_dict: dictionary containing parameters
+        """
+        self.mu = parameter_dict["mu"].to(self.mu)
+        self.set_weights(parameter_dict["weights"])
+        self.set_cov(parameter_dict["cov"])
+
+    def component_log_likelihoods(self, x, cov_regularization=False):
+        r"""
+        returns a tensor containing the log likelihood values of x for each component
+
+        :param torch.Tensor x: input data of shape batch_dimension x dimension
+        :param bool cov_regularization: whether using regularized covariance matrices
+        """
+        if cov_regularization:
+            cov_inv = self._cov_inv_reg
+            logdet_cov = self._logdet_cov_reg
+        else:
+            cov_inv = self._cov_inv
+            logdet_cov = self._logdet_cov
+        centered_x = x[None, :, :] - self.mu[:, None, :]
+        exponent = torch.sum(torch.bmm(centered_x, cov_inv) * centered_x, 2)
+        component_log_likelihoods = (
+            -0.5 * logdet_cov[:, None]
+            - 0.5 * exponent
+            - 0.5 * self.dimension * torch.log(torch.tensor(2 * torch.pi).to(x))
+        )
+        return component_log_likelihoods.T
+
+    def forward(self, x):
+        r"""
+        evaluate negative log likelihood function
+
+        :param torch.Tensor x: input data of shape batch_dimension x dimension
+        """
+        component_log_likelihoods = self.component_log_likelihoods(x)
+        component_log_likelihoods = component_log_likelihoods + torch.log(
+            self._weights[None, :]
+        )
+        log_likelihoods = torch.logsumexp(component_log_likelihoods, -1)
+        return -log_likelihoods
+
+    def classify(self, x, cov_regularization=False):
+        """
+        returns the index of the most likely component
+
+        :param torch.Tensor x: input data of shape batch_dimension x dimension
+        :param bool cov_regularization: whether using regularized covariance matrices
+        """
+        component_log_likelihoods = self.component_log_likelihoods(
+            x, cov_regularization=cov_regularization
+        )
+        component_log_likelihoods = component_log_likelihoods + torch.log(
+            self._weights[None, :]
+        )
+        val, ind = torch.max(component_log_likelihoods, 1)
+        return ind
+
+    def fit(
+        self,
+        dataloader,
+        max_iters=100,
+        stopping_criterion=None,
+        data_init=True,
+        cov_regularization=1e-5,
+        verbose=False,
+    ):
+        """
+        Batched Expectation Maximization algorithm for parameter estimation.
+
+
+        :param torch.utils.data.DataLoader dataloader: containing the data
+        :param int max_iters: maximum number of iterations
+        :param float stopping_criterion: stop when objective decrease is smaller than this number.
+            None for performing exactly max_iters iterations
+        :param bool data_init: True for initialize mu by the first data points, False for using current values as initialization
+        :param bool verbose: Output progress information in the console
+        """
+        if data_init:
+            self.mu = next(iter(dataloader))[0][: self.n_components].to(self.mu)
+
+        objective = 1e100
+        for step in (progress_bar := tqdm(range(max_iters), disable=not verbose)):
+            weights_new, mu_new, cov_new, objective_new = self._EM_step(
+                dataloader, verbose
+            )
+            # stopping criterion
+            self.set_weights = weights_new
+            self.mu = mu_new
+            cov_new_reg = cov_new + cov_regularization * torch.eye(self.dimension)[
+                None, :, :
+            ].tile(self.n_components, 1, 1).to(cov_new)
+            self.set_cov(cov_new_reg)
+            if stopping_criterion:
+                if objective - objective_new < stopping_criterion:
+                    return
+            objective = objective_new
+            progress_bar.set_description(
+                "Step {}, Objective {:.4f}".format(step + 1, objective.item())
+            )
+
+    def _EM_step(self, dataloader, verbose):
+        """
+        one step of the EM algorithm
+
+        :param torch.data.Dataloader dataloader: containing the data
+        :param bool verbose: Output progress information in the console
+        """
+        objective = 0
+        weights_new = torch.zeros_like(self._weights)
+        mu_new = torch.zeros_like(self.mu)
+        C_new = torch.zeros_like(self._cov)
+        n = 0
+        objective = 0
+        for x, _ in tqdm(dataloader, disable=not verbose):
+            x = x.to(self.mu)
+            n += x.shape[0]
+            component_log_likelihoods = self.component_log_likelihoods(x)
+            log_betas = component_log_likelihoods + torch.log(self._weights[None, :])
+            log_beta_sum = torch.logsumexp(log_betas, -1)
+            log_betas = log_betas - log_beta_sum[:, None]
+            objective -= torch.sum(log_beta_sum)
+            betas = torch.exp(log_betas)
+            weights_new += torch.sum(betas, 0)
+            beta_times_x = x[None, :, :] * betas.T[:, :, None]
+            mu_new += torch.sum(beta_times_x, 1)
+            C_new += torch.bmm(
+                beta_times_x.transpose(1, 2),
+                x[None, :, :].tile(self.n_components, 1, 1),
+            )
+        mu_new = mu_new / weights_new[:, None]
+        cov_new = C_new / weights_new[:, None, None] - torch.matmul(
+            mu_new[:, :, None], mu_new[:, None, :]
+        )
+        weights_new = weights_new / n
+        objective = objective / n
+        return weights_new, mu_new, cov_new, objective
